pyproject.toml
[tool.poetry]
name = "djangoproject"
version = "0.1.0"
description = ""
authors = ["huan1894380339 <huan.duong@kyanon.digital>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.2"
djangorestframework = "^3.13.1"
django-environ = "^0.8.1"
psycopg2 = "^2.9.3"
Pillow = "^9.1.0"
django-cors-headers = "^3.12.0"
PyJWT = "^2.3.0"
DateTime = "^4.4"
django-googledrive-storage = "^1.6.0"
PyDrive = "^1.3.1"
django-import-export = "^2.8.0"
tablib = "^3.2.1"
ipdb = "^0.13.9"
pre-commit = "2.19.0"
autopep8 = "^1.6.0"
flake8 = "^4.0.1"
black = "^22.3.0"
add-trailing-comma = "^2.2.3"
isort = "5.8.0"
update = "^0.0.1"
djangorestframework-simplejwt = "^5.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
